+++ R/class-workbook.R
@@ -6353,6 +6353,67 @@ wbWorkbook <- R6::R6Class(
       invisible(self)
     },
 
+    #' @description create dxfs style
+    #' These styles are used with conditional formatting and custom table styles
+    #' @param name the style name
+    #' @param font_name the font name
+    #' @param font_size the font size
+    #' @param font_color the font color (a `wb_color()` object)
+    #' @param numFmt the number format
+    #' @param border logical if borders are applied
+    #' @param border_color the border color
+    #' @param border_style the border style
+    #' @param bgFill any background fill
+    #' @param gradientFill any gradient fill
+    #' @param text_bold logical if text is bold
+    #' @param text_italic logical if text is italic
+    #' @param text_underline logical if text is underlined
+    #' @param ... additional arguments passed to `create_dxfs_style()`
+    #' @return The `wbWorksheetObject`, invisibly
+    #' @export
+    add_dxfs_style = function(
+      name,
+      font_name      = NULL,
+      font_size      = NULL,
+      font_color     = NULL,
+      numFmt         = NULL,
+      border         = NULL,
+      border_color   = wb_color(getOption("openxlsx2.borderColor", "black")),
+      border_style   = getOption("openxlsx2.borderStyle", "thin"),
+      bgFill         = NULL,
+      gradientFill   = NULL,
+      text_bold      = NULL,
+      text_italic    = NULL,
+      text_underline = NULL,
+      ...
+    ) {
+
+      xml_style <- create_dxfs_style(
+        font_name      = font_name,
+        font_size      = font_size,
+        font_color     = font_color,
+        numFmt         = numFmt,
+        border         = border,
+        border_color   = border_color,
+        border_style   = border_style,
+        bgFill         = bgFill,
+        gradientFill   = gradientFill,
+        text_bold      = text_bold,
+        text_italic    = text_italic,
+        text_underline = text_underline,
+        ...            = ...
+      )
+
+      got <- self$styles_mgr$get_dxf_id(name)
+
+      if (!is.null(got) && !is.na(got))
+        warning("dxfs style names should be unique")
+
+      self$add_style(xml_style, name)
+
+      invisible(self)
+    },
+
     #' @description clone style from one sheet to another
     #' @param from the worksheet you are cloning
     #' @param to the worksheet the style is applied to