+++ tests/testthat/test-class-workbook-wrappers.R
@@ -488,7 +488,7 @@ test_that("wb_add_cell_style() is a wrapper", {
   )
 })
 
-# wb_add_named_style() -----------------------------------------------------
+# wb_add_named_style() ----------------------------------------------------
 
 test_that("wb_add_named_style() is a wrapper", {
   wb <- wb_workbook()$add_worksheet(1)
@@ -496,6 +496,17 @@ test_that("wb_add_named_style() is a wrapper", {
     "add_named_style",
     wb = wb,
     params = list(name = "Title")
+  )
+})
+
+# wb_add_dxfs_style() -----------------------------------------------------
+
+test_that("wb_add_dxfs_style() is a wrapper", {
+  wb <- wb_workbook()$add_worksheet(1)
+  expect_wrapper(
+    "add_dxfs_style",
+    wb = wb,
+    params = list(name = "dxfs")
   )
 })